% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeda-functions.R
\name{checkPredictorsAndTree}
\alias{checkPredictorsAndTree}
\title{Check predictors}
\usage{
checkPredictorsAndTree(predictors, tree)
}
\description{
Checks whether the predictors are consistent with the tree
structure.
}
\keyword{internal}
