% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeda-functions.R
\name{coef.treeda}
\alias{coef.treeda}
\title{Get the coefficients from a treeda fit}
\usage{
\method{coef}{treeda}(object, type = c("leaves", "nodes"), ...)
}
\arguments{
\item{object}{An object of class treeda}

\item{type}{Should the coefficients be in the leaf space or the
node space?}

\item{...}{Additional arguments.}
}
\description{
Get the coefficients from a treeda fit
}
