% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeda-functions.R
\name{edgesToChildren}
\alias{edgesToChildren}
\title{Makes a hash table with nodes and their children}
\usage{
edgesToChildren(edges)
}
\description{
Takes the edge matrix from a phylo-class object and turns it into a
list where the entries are nodes and the elements are vectors with
the children of the nodes.
}
\keyword{internal}
