% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{expand_background}
\alias{expand_background}
\title{Expand the background of a gtable.}
\usage{
expand_background(gtable)
}
\arguments{
\item{gtable}{A gtable object whose background needs to be expanded
to fill the whole space.}
}
\value{
A gtable object with a bigger background.
}
\description{
Expand the background of a gtable.
}
\keyword{internal}
