% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{get_leaf_position}
\alias{get_leaf_position}
\title{Get leaf positions from a tree layout}
\usage{
get_leaf_position(tree, ladderize)
}
\arguments{
\item{tree}{A phylogenetic tree}

\item{ladderize}{FALSE for a non-ladderzied layout, TRUE or "right"
for a ladderized layout, "left" for a layout ladderized the other
way.}
}
\description{
Takes a tree, returns a vector with names describing the OTU and
entries giving the position of that OTU in the tree layout.
}
