% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_raster_clip}
\alias{get_raster_clip}
\title{Get raster clip}
\usage{
get_raster_clip(sf, infile, outfile)
}
\arguments{
\item{sf}{An sf object.}

\item{infile}{A file path to the simple feature to clip.}

\item{outfile}{A file path to write the clip to.}
}
\value{

}
\description{
Get raster clip
}
\note{
Sometimes you will need to double quote your path if there are spaces or unique
characters that CLI needs quoted, e.g. '"C:/josh erickson/some file path/"'.
}
