% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_raster_mask}
\alias{get_raster_mask}
\title{Get rast mask}
\usage{
get_raster_mask(sf, outfile, vals, unit = TRUE)
}
\arguments{
\item{sf}{A sf object that has}

\item{outfile}{The path to put the masked .tif files.}

\item{vals}{A gdalinfo call via vapour package.}

\item{unit}{Logical if CHM units are to be converted to feet.}
}
\value{
Writes masked .tif files to the outfile folder.
}
\description{
This function uses terra::mask() to mask the input polygons to the Canopy Height Model (CHM) that
was clipped with the get_raster_clip(). This function is used within a furrr::future_map() call.
}
