% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_raster_stats}
\alias{get_raster_stats}
\title{Get raster stats}
\usage{
get_raster_stats(sf, outfile, vals, FUN = function(x) {
    x * 0.17 + 3
})
}
\arguments{
\item{sf}{A sf object (polygon).}

\item{outfile}{A file path to read-in the .tif files.}

\item{vals}{A gdalinfo call via vapour package.}

\item{FUN}{A function to use as lidR::lmf() function in lidR::find_trees().}
}
\value{
A data.frame with statistics per id.
}
\description{
Get raster stats
}
