% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeda-functions.R
\name{getBranchLengths}
\alias{getBranchLengths}
\title{Make branch length vector}
\usage{
getBranchLengths(tree)
}
\arguments{
\item{tree}{A tree object of class phylo}
}
\value{
A vector of length ntips + nnodes, with ith element giving
the length of the branch above node i.
}
\description{
Gets the branch lengths of the tree, with order the same as the
columns in makeDescendantMatrix.
}
\keyword{internal}
