% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeda-functions.R
\name{makeClassProperties}
\alias{makeClassProperties}
\title{Compute properties of the classes}
\usage{
makeClassProperties(response, projections)
}
\arguments{
\item{response}{A vector containing the response for each observation.}

\item{projections}{A matrix giving the projections of each
observation onto the discriminating axes.}
}
\description{
For each class, computes the prior probabilities, means, and
variances for that class.
}
\keyword{internal}
