% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeda-functions.R
\name{makeDescendantMatrix}
\alias{makeDescendantMatrix}
\title{Make descendant matrix}
\usage{
makeDescendantMatrix(tree)
}
\arguments{
\item{tree}{A tree object of class phylo}
}
\value{
A matrix describing the ancestry structure of a tree.
}
\description{
Make a matrix describing the ancestry structure of a tree. Element
(i,j) indicates whether leaf i is a descendant of node j.
}
\keyword{internal}
