% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeda-functions.R
\name{makeLeafCoefficients}
\alias{makeLeafCoefficients}
\title{Make leaf coefficients}
\usage{
makeLeafCoefficients(sda.out, descendantMatrix, means, sds)
}
\arguments{
\item{sda.out}{A fitted sda object}

\item{descendantMatrix}{A matrix describing the tree which was
used, element (i,j) indicates whether leaf i is a descendant of
node j.}

\item{means}{If the original predictor matrix was centered, the
means of the original predictor matrix, otherwise NULL.}

\item{sds}{If the original predictor matrix was scaled, the sds of
the original predictor matrix, otherwise NULL.}
}
\value{
A list giving the coefficients on the leaves for each of
the discriminating axes and the intercepts for each of the
discriminating axes.
}
\description{
For a set of coefficients defined on a matrix of (potentially
centered and scaled) leaf and node predictors, find the equivalent
set of coefficients on just the leaves.
}
\keyword{internal}
