% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeda-functions.R
\name{makeResponseMatrix}
\alias{makeResponseMatrix}
\title{Make response matrix}
\usage{
makeResponseMatrix(response, class.names = NULL)
}
\arguments{
\item{response}{A factor or character vector containing the
classes.}

\item{class.names}{A character vector giving the possible levels of
the factor. If NULL, it will be generated from the levels of
response.}
}
\value{
A dummy variable matrix with column names giving the class
names.
}
\description{
Create a dummy variable matrix for the response
}
\keyword{internal}
