% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeda-functions.R
\name{nodeToLeafCoefficients}
\alias{nodeToLeafCoefficients}
\title{Node coefficients to leaf coefficients}
\usage{
nodeToLeafCoefficients(coef.vec, tree)
}
\arguments{
\item{coef.vec}{A vector containing coefficients on internal nodes plus leaves.}

\item{tree}{The phylogenetic tree.}
}
\value{
A vector containing coefficients on the leaves.
}
\description{
General-purpose function for going from a coefficient vector on the
nodes to a coefficient vector on the leaves.
}
