% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeda-functions.R
\name{plot.treeda}
\alias{plot.treeda}
\title{Plot treeda object}
\usage{
\method{plot}{treeda}(obj, type = c("both", "variables", "samples"),
  axes = c(1, 2))
}
\arguments{
\item{obj}{The treeda object to plot}

\item{type}{Plot the variables, the samples, or both?}

\item{axes}{The axes to plot the samples along.}
}
\value{
A ggplot object
}
\description{
Provides a plot of the projections of the samples onto the
discriminating axes
}

