% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeda-functions.R
\name{plot.treedacv}
\alias{plot.treedacv}
\title{Plot a treedacv object}
\usage{
\method{plot}{treedacv}(x, ...)
}
\arguments{
\item{x}{An object of class treedacv.}

\item{...}{Additional arguments (not used).}
}
\description{
Gives a plot of the cross-validation error with standard error bars
}
