% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plot_coefficients}
\alias{plot_coefficients}
\title{Plot the discriminating vectors from treeda}
\usage{
plot_coefficients(out.treeda, remove.bl = TRUE, ladderize = TRUE,
  tree.height = 2)
}
\arguments{
\item{out.treeda}{The object resulting from a call to treeda.}

\item{remove.bl}{A logical, TRUE if the tree should be plotted
after setting all branch lengths equal to the same value (tends
to make trees that look nicer) or not.}

\item{ladderize}{Layout parameter for the tree.}
}
\value{
A plot of the tree and the coefficients.
}
\description{
A function that will plot the tree as well as the coefficients
selected for the discriminating axes.
}
