% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plot_tree_and_data}
\alias{plot_tree_and_data}
\title{Plots a tree and data}
\usage{
plot_tree_and_data(tree, data, annotation = NULL, ladderize = TRUE,
  tree.height = 5, scales = "fixed", annotation.color = "firebrick1",
  barsep = 0.02, barheight = 0.03, ...)
}
\arguments{
\item{tree}{A tree of class phylo}

\item{data}{The data to plot, it needs to have row names that are
the same as the tip names of the tree.}

\item{annotation}{A vector of OTU names that should be marked.}

\item{ladderize}{How to lay out the tree, see tree_layout.}

\item{tree.height}{How big the tree is compared to the rest of the plots.}

\item{scales}{Passed to facet_grid, should be either "free_y" or
"fixed".}

\item{annotation.color}{The color for the annotation dots.}

\item{barsep}{The distance of the taxonomy bar above the tree.}

\item{barheight}{The width of the taxonomy bar.}

\item{...}{Additional arguments passed to plot_tree.}
}
\value{
Plots the tree and data, returns a gtable object.
}
\description{
Plots a tree on top and some data associated with the tips on the bottom.
}

