% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeda-functions.R
\name{print.treeda}
\alias{print.treeda}
\title{Print a treeda object}
\usage{
\method{print}{treeda}(x, ...)
}
\arguments{
\item{x}{\code{treeda} object.}

\item{...}{Not used.}
}
\description{
Print a treeda object
}
