% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeda-functions.R
\name{print.treedacv}
\alias{print.treedacv}
\title{Print treedacv object}
\usage{
\method{print}{treedacv}(x, ...)
}
\arguments{
\item{x}{An object of class treedacv}

\item{...}{Optional arguments to be passed to print.treedacv}
}
\description{
Print treedacv object
}
