% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeda-functions.R
\name{coef.treeda}
\alias{coef.treeda}
\title{Get the coefficients from a treeda fit}
\usage{
\method{coef}{treeda}(obj, type = c("leaves", "nodes"))
}
\description{
Get the coefficients from a treeda fit
}

