% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{combine_plot_and_tree}
\alias{combine_plot_and_tree}
\title{Method for combining two ggplots}
\usage{
combine_plot_and_tree(plot, tree.plot, tree.height = 5)
}
\arguments{
\item{plot}{A plot of data about the OTUs with the x axis
corresponding to OTUs.}

\item{treePlot}{A plot of the tree.}

\item{treeHeight}{The relative amount of space in the plot the tree
should take up.}
}
\value{
Prints the combined plot, returns a gtable object.
}
\description{
This method takes a ggplot of some data along the tips of the trees
and a ggplot of a tree and combines them. It assumes that you are
putting the tree on top and that the x axis for the plot has the
OTUs in the correct position (this can be found using the function
get_leaf_position).
}

