% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeda-functions.R
\name{plot.treedacv}
\alias{plot.treedacv}
\title{Plot a treedacv object}
\usage{
\method{plot}{treedacv}(x, ...)
}
\arguments{
\item{x}{An object of class treedacv.}

\item{...}{Not used.}
}
\description{
Plots the cross-validation error with standard error bars.
}
\examples{
data(treeda_example)
out.treedacv = treedacv(response = treeda_example$response,
    predictors = treeda_example$predictors,
    tree = treeda_example$tree,
    pvec = 1:10)
plot(out.treedacv)
}
