% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeda-functions.R
\name{predict.treeda}
\alias{predict.treeda}
\title{Predict using new data}
\usage{
\method{predict}{treeda}(treeda, newdata, newresponse = NULL,
  check.consist = TRUE)
}
\arguments{
\item{treeda}{Output from treeda function}

\item{newdata}{New data}
}
\value{
A list containing the projections of the new data onto the
discriminating axes, the predicted classes, and the rss (if the
ground truth for the responses is available).
}
\description{
Given a fitted treeda model, get the predicted classes and
projections onto the discriminating axes for new data.
}
