% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeda-functions.R
\name{print.treedacv}
\alias{print.treedacv}
\title{Print treedacv objects}
\usage{
\method{print}{treedacv}(x, ...)
}
\arguments{
\item{x}{\code{treedacv} object.}

\item{...}{Not used}
}
\description{
Print treedacv objects
}
