% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeda-functions.R
\name{treeda}
\alias{treeda}
\title{Tree-based SDA}
\usage{
treeda(response, predictors, tree, p, k = nclasses - 1, center = TRUE,
  scale = TRUE, class.names = NULL, check.consist = TRUE, A = NULL, ...)
}
\arguments{
\item{response}{A factor or character vector giving the class to be
predicted.}

\item{predictors}{A matrix of predictor variables corresponding to
the leaves of the tree and in the same order as the leaves of the
tree.}

\item{tree}{A tree of class phylo.}

\item{p}{The number of predictors to use.}

\item{k}{The number of components to use.}

\item{...}{Additional arguments to be passed to sda}
}
\description{
Performs sparse discriminant analysis on tree-structured data by
augmenting the predictor matrix with additional predictors
corresponding to the nodes.
}

